+++ app/room/page.tsx
@@ -15,8 +15,8 @@ import { useEffect, useState } from 'react';
 
 export default function Page() {
   // TODO: get user input for room and name
-  const room = "quickstart-room";
-  const name = "quickstart-user";
+  const room = "beta-tester-room";
+  const name = "beta-tester";
   const [token, setToken] = useState("");
 
   useEffect(() => {