+++ lib/db.ts
@@ -5,7 +5,6 @@ declare global  {
 };
 
 export const db = globalThis.prisma  || new PrismaClient();
-// export const db = globalThis.prisma  || new PrismaClient();
 
 // if(process.env.NODE_ENV !=="production") 
     globalThis.prisma = db;